backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers
pydantic
pdfplumber
python-docx
pytesseract
openai
python-multipart
PyMuPDF
python-dotenv
